a workable version with parallel tests of other compressors on Andes

+++ include/Retriever/HPSSFileRetriever.hpp
@@ -65,6 +65,7 @@ namespace MDR {
 			//adios2::IO readIO = ad.DeclareIO(filename);
 			MPI_Barrier(MPI_COMM_WORLD);
                         adios2::Engine bpFileReader = readIO.Open(filename, adios2::Mode::Read);
+			/*
 			if(bpFileReader){
 				if(!rank){
 					const std::map<std::string, adios2::Params> variables =
@@ -86,8 +87,9 @@ namespace MDR {
 				std::cout << rank << " " << size << std::endl;
 				MPI_Abort(MPI_COMM_WORLD, -1);
 			}
-
+			*/
                         adios2::Variable<uint8_t> bp_fdata = readIO.InquireVariable<uint8_t>(filename);
+			/*
 			if(bp_fdata){
 				std::cout << filename << "_found in rank " << rank << "\n";
 				fflush(stdout);
@@ -109,7 +111,8 @@ namespace MDR {
 				std::cout << "list of existing variables done\n";
 				MPI_Abort(MPI_COMM_WORLD, 0);
 			}
-                        bp_fdata.SetSelection(adios2::Box<adios2::Dims>({level_segment_size[i][segment_offsets[i]] * rank}, {level_segment_size[i][segment_offsets[i]]}));
+			*/
+                        bp_fdata.SetSelection(adios2::Box<adios2::Dims>({(uint64_t)level_segment_size[i][segment_offsets[i]] * rank}, {(uint64_t)level_segment_size[i][segment_offsets[i]]}));
                         bpFileReader.Get<uint8_t>(bp_fdata, buffer, adios2::Mode::Sync);
                         bpFileReader.Close();
                     }
